Also cache blob and url in JSONWriter
in addition to stringified

Don't generate new object url and waste memory if it hasn't yet been revoked.

Slight difference in behavior: the revoke function outputted from toURL is always the same thing. it revokes whatever the current url is; not the url that was outputted with the revoke function.